Refactor main.go
