+++ cmd/version.go
@@ -1,0 +1,21 @@
+package cmd
+
+import "github.com/spf13/cobra"
+
+var version = "dev"
+
+func getVersionCmd() *cobra.Command {
+	return &cobra.Command{
+		Use:   "version",
+		Short: "Prints the version of the binary",
+		RunE: func(cmd *cobra.Command, args []string) error {
+			return printVersion(cmd, args)
+		},
+	}
+}
+
+func printVersion(cmd *cobra.Command, args []string) error {
+	println(version)
+
+	return nil
+}
