+++ main.go
@@ -3,26 +3,7 @@ package main
 import (
 	"github.com/cheqd/did-resolver/cmd"
 	_ "github.com/cheqd/did-resolver/docs"
-	"github.com/spf13/cobra"
 )
-
-var version = "dev"
-
-func getVersionCmd() *cobra.Command {
-	return &cobra.Command{
-		Use:   "version",
-		Short: "Prints the version of the binary",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			return printVersion(cmd, args)
-		},
-	}
-}
-
-func printVersion(cmd *cobra.Command, args []string) error {
-	println(version)
-
-	return nil
-}
 
 // @title DID Resolver for did:cheqd method
 // @version 1.0
@@ -36,7 +17,6 @@ func printVersion(cmd *cobra.Command, args []string) error {
 // @schemes	https http
 func main() {
 	rootCmd := cmd.GetRootCmd()
-	rootCmd.AddCommand(getVersionCmd())
 
 	if err := rootCmd.Execute(); err != nil {
 		panic(err)
